refactor: 수정 요청 DTO 이름 정리
- API 요청 타입에서 Dto 접미사를 제거해 기존 요청 DTO 네이밍과 맞춤
- 같은 이름의 저장 모델과 충돌하지 않도록 엔티티 타입을 명시적으로 분리
- 서비스와 테스트 참조를 새 타입명에 맞춰 정리해 FQCN 없이 읽히도록 유지
