+++ src/main/java/gg/agit/konect/domain/club/repository/ClubInformationUpdateRequestRepository.java
@@ -2,8 +2,8 @@ package gg.agit.konect.domain.club.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import gg.agit.konect.domain.club.model.ClubInformationUpdateRequest;
+import gg.agit.konect.domain.club.model.ClubInformationUpdateRequestEntity;
 
 public interface ClubInformationUpdateRequestRepository
-    extends JpaRepository<ClubInformationUpdateRequest, Integer> {
+    extends JpaRepository<ClubInformationUpdateRequestEntity, Integer> {
 }
